Fix for audio volume, js

+++ main/snd_c3.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 
-typedef void (snd_cb_t)(int8_t *buf, int buflen);
+typedef void (snd_cb_t)(int16_t *buf, int buflen);
 void snd_init(int samprate, snd_cb_t *cb);
